misc: addressed feedback
